Add new skip_prefix argument to build_store_tables_data_processor
allowing to specify a list of prefix text that should not be stored in
intermediary table data structure.

This is intended to be used in place of build_skip_block_starting_with_processor
to remove undesired text, since the latest may skip undesired parts when
checking on a parent object (e.g. TextBox) that my group semantically unrelated
text.